Sanctuary/requirements.txt
discord.py=>1.0.0
asyncpg
